webui/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-bcrypt>=1.0
PyJWT>=2.8
